+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.1
+
+* fix 'missing-field' param issue @phoet https://github.com/yurijmi/omniauth-telegram/pull/7
+* fix 'invalid_signature' with missing username issue @phoet https://github.com/yurijmi/omniauth-telegram/pull/7
+
 ## 0.2.0
 
 * update gem requirements for compatibility with OmniAuth 2.0 @phoet https://github.com/yurijmi/omniauth-telegram/pull/7